backend_services/microservice_1/requirements.txt
# Allgemeine Abhängigkeiten
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2
sqlalchemy==2.0.23
python-dotenv==1.0.0
pyyaml==6.0.1

# ASGI-Server und -Tools
starlette==0.27.0
httptools==0.6.1
websockets==12.0

# Typing Extensions
typing-extensions==4.8.0

# Async Tools
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
greenlet==3.0.1

# HTTP und Utilities
click==8.1.7
colorama==0.4.6
h11==0.14.0
idna==3.5
sniffio==1.3.0
watchfiles==0.21.0

# Pip tools
pip==23.3.1
setuptools==68.0.0
wheel==0.41.2

# SQLAlchemy für ORM
sqlalchemy==2.0.23

# Datenbanktreiber, falls noch nicht hinzugefügt
psycopg2-binary==2.9.3

# Für das Passwortverschleicher
bcrypt==4.1.1

pydantic==2.5.2  
pydantic-core==2.14.5
email-validator>=1.1.3 
annotated-types>=0.4.0 
typing-extensions>=4.6.1 